+++ src/index.ts
@@ -1,8 +1,4 @@
 import {registerPlugin} from 'kaltura-player-js';
-import {pluginName, UIManagersPlugin} from './services-managers';
+import {pluginName, UIManagersPlugin} from './services-manager';
 
 registerPlugin(pluginName, UIManagersPlugin);
-
-export * from './services/side-panels/types/types';
-export {SidePanelsManager} from './services/side-panels/side-panels-manager';
-export {ISidePanelItemDto} from './services/side-panels/models/side-panel-item-dto';
