+++ types/index.d.ts
@@ -1,3 +1,6 @@
-export * from './services/side-panels/types/types';
-export { SidePanelsManager } from './services/side-panels/side-panels-manager';
-export { ISidePanelItemDto } from './services/side-panels/models/side-panel-item-dto';
+/// <reference types="./modules" />
+
+export { pluginName } from 'services-manager';
+export * from "services/side-panels/types/types";
+export { SidePanelsManager } from "services/side-panels/side-panels-manager";
+export { ISidePanelItemDto } from "services/side-panels/models/side-panel-item-dto";
