--- types/services-managers.d.ts
@@ -1,1 +1,0 @@
-export declare const pluginName = "uiManagers";
