--- types/services/side-panels/models/item-wrapper.d.ts
@@ -1,1 +1,0 @@
-export {};
