--- types/services/side-panels/ui/side-panel.component.d.ts
@@ -1,1 +1,0 @@
-export {};
